# utils/swagger-utils.yml
openapi: 3.0.3
info:
  title: Localhost
  version: 1.0.0
  contact: {}
servers:
  - url: http://localhost
paths:
...
  /createOrders:
    post:
      tags:
        - Orders
      summary: createOrders
      description: createOrders
      operationId: createorders
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cod_produto:
                  type: number
                  example: 1
            example:
              cod_produto: 1
      responses:
        '200':
          description: ''
...
  /updatepOrders:
    put:
      tags:
        - Orders
      summary: updateOrders
      description: updateOrders
      operationId: updateorders
      parameters:
        - name: id
          in: query
          schema:
            type: string
            example: '1'
          description: INT
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cod_produto:
                  type: number
                  example: 1
                qtde_pedido:
                  type: string
                  example: ''
            example:
              cod_produto: 1
              qtde_pedido: ''
      responses:
        '200':
          description: ''
  /deleteOrders:
    delete:
      tags:
        - Orders
      summary: deleteOrders
      description: deleteOrders
      operationId: deleteorders
      parameters:
        - name: id
          in: query
          schema:
            type: string
            example: '1'
          description: INT
      responses:
        '200':
          description: ''
...